some graphic functions

+++ plogic.h
@@ -5,7 +5,7 @@
 #include "common.h"
 
 void swappoints(struct point *points, short int a, short int b);
-short int turn(const struct point a, const struct point b, const struct point c);
+long int turn(const struct point a, const struct point b, const struct point c);
 short int checklength(const struct point current, const struct point a, const struct point b);
 short int findpoints(struct point *points, const short int pointsnum);
